Set block literal struct defintions to their actual size
The size of a block literal is known because it can be found in the block descriptor. It helps the decompiler and user understand whats going on better by making it clear a stack variable assignment is being stored in the block literal and not just an unrelated stack variable assignment. Fixes issue #1